Restructured project files; Added uniform access from shader

+++ ArchEngine/util/Timer.cpp
@@ -14,7 +14,6 @@ namespace arch {
 	void Timer::Update() {
 		last = now;
 		now = SDL_GetPerformanceCounter();
-		//deltaTime = (double)((NOW - LAST)*1000 / (double)SDL_GetPerformanceFrequency() );
-		deltaTime = static_cast<double>( (double)(now - last) * 1000 / (double)SDL_GetPerformanceFrequency()) * 0.001f;
+		deltaTime = static_cast<float>( (float)(now - last) * 1000 / (float)SDL_GetPerformanceFrequency()) * 0.001f;
 	}
 }
